# src/api/openapi.yaml
openapi: 3.0.0
info:
  title: Autonomous Content Creation Service API
  description: |
    API for the Autonomous Content Creation Service, providing endpoints for client management,
    project management, content creation, and financial operations.
  version: 1.0.0
  contact:
    name: Autonomous Content Service
    url: https://github.com/Ceesaxp/autonomous-content-service
servers:
  - url: /api/v1
    description: API server
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      summary: List projects
      description: Retrieves a list of projects with optional filtering
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ClientIdQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/Pagination'
      responses:
        '200':
          description: Projects retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProjectSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
    
    post:
      summary: Create a new project
      description: Creates a new content creation project
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{projectId}:
    get:
      summary: Get project details
      description: Retrieves detailed information about a specific project
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      summary: Update project details
      description: Updates information for a specific project
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    delete:
      summary: Cancel a project
      description: Cancels a specific project (does not delete it completely)
      tags:
        - Projects
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: Project cancelled successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/content:
    get:
      summary: List content for a project
      description: Retrieves a list of content items for a specific project
      tags:
        - Content
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/ContentStatusQuery'
      responses:
        '200':
          description: Content list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    post:
      summary: Request new content
      description: Requests a new content item for a specific project
      tags:
        - Content
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentRequest'
      responses:
        '201':
          description: Content requested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/{contentId}:
    get:
      summary: Get content details
      description: Retrieves detailed information about a specific content item
      tags:
        - Content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    
    put:
      summary: Update content
      description: Updates a specific content item
      tags:
        - Content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentUpdate'
      responses:
        '200':
          description: Content updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/{contentId}/versions:
    get:
      summary: Get content version history
      description: Retrieves version history for a specific content item
      tags:
        - Content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content versions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContentVersion'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /content/{contentId}/approve:
    post:
      summary: Approve content
      description: Approves a specific content item
      tags:
        - Content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Content'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Content cannot be approved in its current state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      description: ID of the project
      schema:
        type: string
        format: uuid
    ContentId:
      name: contentId
      in: path
      required: true
      description: ID of the content item
      schema:
        type: string
        format: uuid
...
    ClientIdQuery:
      name: clientId
      in: query
      required: false
      description: Filter by client ID
      schema:
        type: string
        format: uuid
...
    StatusQuery:
      name: status
      in: query
      required: false
      description: Filter by status
      schema:
        type: string
    ContentStatusQuery:
      name: status
      in: query
      required: false
      description: Filter by content status
      schema:
        type: string
        enum: [Planning, Researching, Drafting, Editing, Review, Approved, Published, Archived]
    Pagination:
      name: page
      in: query
      required: false
      description: Page number (starting from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
      name: limit
      in: query
      required: false
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    Pagination:
      type: object
      properties:
        totalItems:
          type: integer
        totalPages:
          type: integer
        currentPage:
          type: integer
        itemsPerPage:
          type: integer
...
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: float
        currency:
          type: string
          minLength: 3
          maxLength: 3
...
    ProjectCreate:
      type: object
      required:
        - title
        - description
        - contentType
        - deadline
        - budget
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
        contentType:
          type: string
          enum: [BlogPost, SocialPost, EmailNewsletter, WebsiteCopy, TechnicalArticle, ProductDescription, PressRelease]
        deadline:
          type: string
          format: date-time
        budget:
          $ref: '#/components/schemas/Money'
        priority:
          type: string
          enum: [High, Medium, Low]
          default: Medium
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
        deadline:
          type: string
          format: date-time
        budget:
          $ref: '#/components/schemas/Money'
        priority:
          type: string
          enum: [High, Medium, Low]
        status:
          type: string
          enum: [Draft, Planning, InProgress, Review, Completed, Cancelled]
    ProjectSummary:
      type: object
      required:
        - projectId
        - clientId
        - title
        - contentType
        - deadline
        - status
        - createdAt
      properties:
        projectId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        title:
          type: string
        contentType:
          type: string
          enum: [BlogPost, SocialPost, EmailNewsletter, WebsiteCopy, TechnicalArticle, ProductDescription, PressRelease]
        deadline:
          type: string
          format: date-time
        status:
          type: string
          enum: [Draft, Planning, InProgress, Review, Completed, Cancelled]
        createdAt:
          type: string
          format: date-time
    Project:
      type: object
      required:
        - projectId
        - clientId
        - title
        - description
        - contentType
        - deadline
        - budget
        - priority
        - status
        - createdAt
        - updatedAt
      properties:
        projectId:
          type: string
          format: uuid
        clientId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        contentType:
          type: string
          enum: [BlogPost, SocialPost, EmailNewsletter, WebsiteCopy, TechnicalArticle, ProductDescription, PressRelease]
        deadline:
          type: string
          format: date-time
        budget:
          $ref: '#/components/schemas/Money'
        priority:
          type: string
          enum: [High, Medium, Low]
        status:
          type: string
          enum: [Draft, Planning, InProgress, Review, Completed, Cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        content:
          type: array
          items:
            $ref: '#/components/schemas/ContentSummary'
    ContentRequest:
      type: object
      required:
        - title
        - type
      properties:
        title:
          type: string
        type:
          type: string
          enum: [BlogPost, SocialPost, EmailNewsletter, WebsiteCopy, TechnicalArticle, ProductDescription, PressRelease]
        metadata:
          type: object
          description: Additional metadata for content creation
    ContentUpdate:
      type: object
      properties:
        title:
          type: string
        data:
          type: string
        metadata:
          type: object
        status:
          type: string
          enum: [Planning, Researching, Drafting, Editing, Review, Approved, Published, Archived]
    ContentSummary:
      type: object
      required:
        - contentId
        - projectId
        - title
        - type
        - status
        - version
        - wordCount
        - createdAt
      properties:
        contentId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [BlogPost, SocialPost, EmailNewsletter, WebsiteCopy, TechnicalArticle, ProductDescription, PressRelease]
        status:
          type: string
          enum: [Planning, Researching, Drafting, Editing, Review, Approved, Published, Archived]
        version:
          type: integer
        wordCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    Content:
      type: object
      required:
        - contentId
        - projectId
        - title
        - type
        - status
        - version
        - wordCount
        - createdAt
        - updatedAt
      properties:
        contentId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        type:
          type: string
          enum: [BlogPost, SocialPost, EmailNewsletter, WebsiteCopy, TechnicalArticle, ProductDescription, PressRelease]
        status:
          type: string
          enum: [Planning, Researching, Drafting, Editing, Review, Approved, Published, Archived]
        data:
          type: string
        metadata:
          type: object
        version:
          type: integer
        wordCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        statistics:
          type: object
          properties:
            readabilityScore:
              type: number
              format: float
            seoScore:
              type: number
              format: float
            engagementScore:
              type: number
              format: float
            plagiarismScore:
              type: number
              format: float
    ContentVersion:
      type: object
      required:
        - versionId
        - contentId
        - versionNumber
        - data
        - createdAt
        - createdBy
      properties:
        versionId:
          type: string
          format: uuid
        contentId:
          type: string
          format: uuid
        versionNumber:
          type: integer
        data:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
...